Remove user consent class (moved to next-consent-proxy)
 🐿 v2.8.0
