+++ README.md
@@ -123,88 +123,6 @@ appName - the name of the app using `n-user-api-client`
 
 If successful, the user will be reauthenticated and the [fresh session data](https://developer.ft.com/portal/docs-membership-platform-api-post-login) will be returned.
 
-## Classes
-
-### UserConsent
-
-By default, requires the `MEMBERSHIP_API_HOST_PROD` and `MEMBERSHIP_API_KEY_PROD` environment variables, respectively `..._TEST` / `..._MOCK` (see `mode` below). If the optional `options` argument contains an `envPrefix` property, these environment variables used will be `${envPrefix}_HOST_*` and `${envPrefix}_KEY_*`.
-
-```js
-import { UserConsent } from '@financial-times/n-user-api-client';
-
-// optional overrides
-const options = {
-  envPrefix: 'MY_API', // use MY_API_HOST_* and MY_API_KEY_* env vars, defaults to MEMBERSHIP_API
-  apiKeyHeader: 'ft-api-key', // override the api key header name, defaults to x-api-key
-  path: '/my/path' // override the base path after the host, defaults to /consent/users/
-}
-
-const api = new UserConsent(uuid, source[, mode = 'PROD', scope = 'FTPINK', options]);
-
-// e.g const api = new UserConsent('user-id-here', 'signup-app');
-```
-
-#### Consent methods
-
-The following methods return and/or accept a `consent unit`, as below:
-
-```js
-const consentPayload = {
-  status: true,
-  fow: "fow-id",
-  lbi: false, // optional, defaults to false
-  source: "my-source" // optional, will overwrite the instance source
-}
-```
-
-Methods available:
-
-```js
-api.getConsent(category, channel);
-
-api.createConsent(category, channel, consentPayload);
-
-api.updateConsent(category, channel, consentPayload);
-```
-
-#### Consent record methods
-
-The following methods return and/or accept a `consent record`, as below:
-
-```js
-const consentRecordPayload = {
-  categoryName1: {
-    channelName1: {
-      status: true,
-      fow: "fow-id",
-      lbi: false, // optional, defaults to false
-      source: "my-source"
-    },
-    channelName2: {
-      status: true,
-      fow: "fow-id",
-      lbi: false, // optional, defaults to false
-      source: "my-source"
-    }
-  },
-  categoryName2: { ... }
-}
-```
-
-Methods available:
-
-```js
-api.getConsentRecord();
-
-api.createConsentRecord(consentRecordPayload);
-// will error if the user has an existing consent record on the scope
-
-api.updateConsentRecord(consentRecordPayload);
-// will perform a deep merge on the existing consent record on the scope
-```
-
-All write requests to the Single Consent Store (consent units / records) have payload validation.
-
 ## Build
 
 The module is written in typescript - compile to the dist/ folder with:
