+++ src/main.ts
@@ -1,7 +1,7 @@
 export { getUserBySession, getUserIdAndSessionData } from './read/getUser';
 export { getPaymentDetailsBySession } from './read/getPayment';
 export { updateUserProfile, changeUserPassword } from './write/updateUser';
-export { UserConsent, ConsentValidator } from './services/user-consent';
+export { ConsentValidator } from './validation/consent-api';
 export { loginUser } from './write/loginUser';
 export { errorTypes, ErrorWithData } from './utils/error';
 export { logger } from './utils/logger';
